Separate "read more" link from post excerpt
The logic that decides whether a post needs a "read more" link or not
has moved out of the model and into a template. This seems more
appropriate, since it deals with an issue of display. Post excerpts can
now be gotten absent of any link at the end, which makes them more
useful for use with meta description tags. The previous setting which
allowed the user to determine the text of the link is no longer needed.
The user can now simply override the template.