Category and tag detail views now support pagination
The trouble with these two views is that they were detail views of
Categories and Tags. Since the main object was the single category or
tag, there was nothing for the generic view to paginate by default.
Initially I implemented my own pagination on the Post queryset that I was
adding to the context, but it seemed simpler just to change both the
views to ListViews of Posts, filtered by Category and Tag.