Simplify post excerpts
The post excerpt property always returns a value. If the post has a
tease, that is returned as the excerpt. Otherwise, the truncated body is
returned.

The only relevant setting is BLOG_EXCERPTLENGTH, which determines how
many words the excerpt should include when the body is being truncated.

Whether the full post body, the tease, or the excerpt is displayed on
post listings and feeds is determined by templates. These are easily
overwritten by the user.

The default behaviour for feeds is to print the tease if it exists,
otherwise print the full body. For post listings it is the same.

Excerpts, then, are only used by default to fill the meta description
tag on a post page if the post has no tease.