Prevent caching of publish date
When passing `now()`, the function is evaluated when the server starts
and is cached. This causes the default publish date to always be
whenever the server was last started. When passing in `now` (without
paranthesis), the function itself is passed in. The function is
evaluated whenever the model is called, which causes the default publish
date to always be the current date.